✨ feat(acceptance): restore in-app workspace and assemble the public page from atoms
Keep the full owner workspace in the main SPA and compose the workbench share page from read-only capability atoms so owner actions stay out of the public bundle.
