+++ apps/workbench/app/routes/acceptanceIndex.tsx
@@ -1,0 +1,1 @@
+export { default, loader } from './homeRedirect';
