--- apps/workbench/app/routes/acceptanceList.tsx
@@ -1,1 +1,0 @@
-export { default } from '../../src/features/acceptance/AcceptanceList';
