--- apps/workbench/src/routes/acceptance/index.tsx
@@ -1,3 +1,0 @@
-'use client';
-
-export { default } from '../../features/acceptance/AcceptanceList';
