+++ src/routes/(main)/acceptance/empty.tsx
@@ -1,0 +1,3 @@
+'use client';
+
+export { AcceptanceEmptyDetail as default } from '@/features/Verify';
