+++ src/routes/(main)/acceptance/index.tsx
@@ -1,0 +1,3 @@
+'use client';
+
+export { AcceptanceWorkspace as default } from '@/features/Verify';
